+++ python_modules/dagster/dagster/core/execution/plan/compute.py
@@ -7,7 +7,7 @@ from dagster.core.execution.context.system import SystemComputeExecutionContext
 from .objects import ExecutionStep, StepInput, StepKind, StepOutput
 
 
-def create_compute_step(pipeline_name, environment_config, solid, step_inputs, handle):
+def create_compute_step(pipeline_name, solid, step_inputs, handle):
     check.str_param(pipeline_name, 'pipeline_name')
     check.inst_param(solid, 'solid', Solid)
     check.list_param(step_inputs, 'step_inputs', of_type=StepInput)
@@ -28,9 +28,7 @@ def create_compute_step(pipeline_name, environment_config, solid, step_inputs, h
         ),
         kind=StepKind.COMPUTE,
         solid_handle=handle,
-        metadata=solid.step_metadata_fn(environment_config)
-        if solid.step_metadata_fn
-        else solid.metadata,
+        metadata=solid.metadata,
     )