+++ CHANGES.md
@@ -53,12 +53,15 @@
 - `Field` takes a `is_required` rather than a `is_optional` argument. This is avoid confusion
   with python's typing and dagster's definition of `Optional`, which indicates None-ability,
   rather than existence. `is_optional` is deprecated and will be removed in a future version.
+- `step_metadata_fn` has been removed from `SolidDefinition` & `@solid`.
 - Dagster Type System Changes
+
   - `RuntimeType` has been renamed to `DagsterType` is now an encouraged API for type creation.
   - Core type check function of DagsterType can now return a naked `bool` in addition
     to a `TypeCheck` object.
   - `define_python_dagster_type` and `dagster_type` no longer take a `type_check` argument. If
     a custom type_check is needed, use `DagsterType`.
+
 **New**
 
 - `dagster/priority` tags can now be used to prioritize the order of execution for the built in in process and multiprocess engines.